met repository and service van genres

+++ src/main/java/be/vdab/retrovideo/entities/Genre.java
@@ -4,7 +4,7 @@ public class Genre {
 	private final long id;
 	private final String naam;
 	
-	public Genre(int id, String naam) {
+	public Genre(long id, String naam) {
 		this.id = id;
 		this.naam = naam;
 	}
